// screens/viewMedia/ViewMedia.types.ts
import { ImageStyle, TextStyle, ViewStyle } from "react-native";

export interface IViewMediaStyles {
  contanier: ViewStyle;
  mainbar: ViewStyle;
  nowPlayingText: TextStyle;
  musicLogoView: ViewStyle;
  imageView: ImageStyle;
  nameOfSongView: ViewStyle;
  nameOfSongText1: TextStyle;
  nameOfSongText2: TextStyle;
  sliderView: ViewStyle;
  sliderStyle: ViewStyle;
  sliderTime: ViewStyle;
  functionsView: ViewStyle;
}